pyproject.toml
[project]
name = "ytresearch"
version = "0.1.0"
description = "YouTube music research archiver — downloads, analyzes, and tags audio from YouTube"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic",
    "mutagen",
    "python-dotenv",
    "yt-dlp",
    "youtube-comment-downloader",
]

[project.scripts]
ytresearch = "ytresearch.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ytresearch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest",
]
